amp-embedly: Implement embedly service.
